+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added  
+
+- Published documentation for `CollectionTree`.
+
 ## [0.0.2] - 2025-02-12
 
 ### Added